spotchecks button fixes
