--- src/NavigationHandler.d.ts
@@ -1,5 +1,0 @@
-import type { SpotcheckState } from './SpotCheckState';
-
-export declare function useSpotcheckNavigation(
-  spotcheckRef: React.MutableRefObject<SpotcheckState>
-): void;
