+++ src/SpotCheck.tsx
@@ -74,6 +74,7 @@ export const initializeSpotChecks = ({
   variables = {},
   customProperties = {},
   listener = spotchecksListener,
+  useSegmentsHook,
 }: SpotcheckProps) => {
   store.dispatch(
     updateState({
@@ -83,6 +84,7 @@ export const initializeSpotChecks = ({
       variables,
       customProperties,
       listener,
+      useSegmentsHook,
     })
   );
 };
