+++ src/SpotCheckState.tsx
@@ -48,6 +48,7 @@ export interface SpotcheckState {
   screenName: string;
   isChat: boolean;
   appearance: Record<string, any>;
+  useSegmentsHook: (() => string[]) | undefined;
 }
 
 const initialState: SpotcheckState = {
@@ -96,6 +97,7 @@ const initialState: SpotcheckState = {
   screenName: '',
   isChat: false,
   appearance: {},
+  useSegmentsHook: undefined,
 };
 
 const spotcheckSlice = createSlice({
