+++ src/Types.tsx
@@ -23,6 +23,7 @@ export interface SpotcheckProps {
   variables?: Variables;
   customProperties?: CustomProperties;
   listener?: SsSpotcheckListener;
+  useSegmentsHook?: () => string[];
 }
 
 export interface TrackEventProps {
