+++ tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig",
-  "exclude": ["example", "**/*.expo.ts", "**/*.expo.tsx"]
+  "exclude": ["example"]
 }
